use 32bit reference counter
MSVC STL uses long.
https://github.com/microsoft/STL/blob/009881898c2112f92ab7ee743eadb148c01329c9/stl/inc/memory#L1029
https://github.com/microsoft/STL/blob/009881898c2112f92ab7ee743eadb148c01329c9/stl/inc/xatomic.h#L103

GNU libstdc++ uses int.
https://github.com/gcc-mirror/gcc/blob/4c57e57b370333ae30feba624ab9caa69a33e15c/libstdc++-v3/include/bits/shared_ptr_base.h#L237

LLVM libc++ uses long.
https://github.com/llvm/llvm-project/blob/da2366d5322ec2c5b95056e0b48c3bc6fcee02bc/libcxx/include/__memory/shared_ptr.h#L137

+++ test/test.cpp
@@ -749,8 +749,8 @@ TEST_CASE("alignment")
 
     struct placeholder
     {
-        std::atomic<long> strong_ref_counter;
-        std::atomic<long> weak_ref_counter;
+        std::atomic<unsigned> strong_ref_counter;
+        std::atomic<unsigned> weak_ref_counter;
         void* vptr;
     };
